Add support for legacy Ubuntu 20.04 bootloader

+++ ubuntu-autoinstall/replace-grub.sh
@@ -7,6 +7,7 @@ set -x
 SOURCE_ISO="ubuntu-22.04.5-live-server-amd64.iso"
 DESTINATION_ISO="ubuntu-autoinstall.iso"
 AUTOINSTALL_CONFIG_FILE="autoinstall.example"
+ISOLINUX_BOOTLOADER=0
 
 error() {
   echo "ERROR $*" >&2
@@ -28,6 +29,7 @@ Replace grub.cfg on Ubuntu ISO
   -d    Destination ISO path
   -g    Grub.cfg file
   -l    Loopback.cfg file
+  -i    Use legacy isolinux bootloader (for Ubuntu 20.04)
 EOF
 }
 
@@ -55,6 +57,9 @@ args() {
         usage
         exit 0
         ;; 
+      (-i | --isolinux)
+        ISOLINUX_BOOTLOADER=1
+        ;;
     esac
     shift
   done
@@ -84,15 +89,18 @@ extract_iso() {
 
   echo "Extracting ISO image..."
   xorriso -osirrox on -indev "${SOURCE_ISO}" -extract / "${ISO_FILESYSTEM_DIR}"
-
+  
   TODAY=$(date +"%Y-%m-%d")
-  MBR_IMAGE="ubuntu-original-$TODAY.mbr"
-  EFI_IMAGE="ubuntu-original-$TODAY.efi"
-  dd if="${SOURCE_ISO}" bs=1 count=446 of="${BOOT_IMAGE_DIR}/${MBR_IMAGE}"
 
-  START_BLOCK=$(fdisk -l "${SOURCE_ISO}" | fgrep '.iso2 ' | awk '{print $2}')
-  SECTORS=$(fdisk -l "${SOURCE_ISO}" | fgrep '.iso2 ' | awk '{print $4}')
-  dd if="${SOURCE_ISO}" bs=512 skip="${START_BLOCK}" count="${SECTORS}" of="${BOOT_IMAGE_DIR}/${EFI_IMAGE}"
+  if [ ${ISOLINUX_BOOTLOADER} -eq 0 ]; then
+    MBR_IMAGE="ubuntu-original-$TODAY.mbr"
+    EFI_IMAGE="ubuntu-original-$TODAY.efi"
+    dd if="${SOURCE_ISO}" bs=1 count=446 of="${BOOT_IMAGE_DIR}/${MBR_IMAGE}"
+
+    START_BLOCK=$(fdisk -l "${SOURCE_ISO}" | fgrep '.iso2 ' | awk '{print $2}')
+    SECTORS=$(fdisk -l "${SOURCE_ISO}" | fgrep '.iso2 ' | awk '{print $4}')
+    dd if="${SOURCE_ISO}" bs=512 skip="${START_BLOCK}" count="${SECTORS}" of="${BOOT_IMAGE_DIR}/${EFI_IMAGE}"
+  fi
 }
 
 configure_grub() {
@@ -108,21 +116,37 @@ configure_grub() {
 }
 
 reassemble_iso() {
-  xorriso -as mkisofs \
-    -r -V "ubuntu-autoinstall-${TODAY}" -J -joliet-long -l \
-    -iso-level 3 \
-    -partition_offset 16 \
-    --grub2-mbr "${BOOT_IMAGE_DIR}/${MBR_IMAGE}" \
-    --mbr-force-bootable \
-    -append_partition 2 0xEF "${BOOT_IMAGE_DIR}/${EFI_IMAGE}" \
-    -appended_part_as_gpt \
-    -c boot.catalog \
-    -b boot/grub/i386-pc/eltorito.img \
-    -no-emul-boot -boot-load-size 4 -boot-info-table --grub2-boot-info \
-    -eltorito-alt-boot \
-    -e '--interval:appended_partition_2:all::' \
-    -no-emul-boot \
-    -o "${DESTINATION_ISO}" "${ISO_FILESYSTEM_DIR}"
+  if [ ${ISOLINUX_BOOTLOADER} -eq 0 ]; then
+    xorriso -as mkisofs \
+      -r -V "ubuntu-autoinstall-${TODAY}" -J -joliet-long -l \
+      -iso-level 3 \
+      -partition_offset 16 \
+      --grub2-mbr "${BOOT_IMAGE_DIR}/${MBR_IMAGE}" \
+      --mbr-force-bootable \
+      -append_partition 2 0xEF "${BOOT_IMAGE_DIR}/${EFI_IMAGE}" \
+      -appended_part_as_gpt \
+      -c boot.catalog \
+      -b boot/grub/i386-pc/eltorito.img \
+      -no-emul-boot -boot-load-size 4 -boot-info-table --grub2-boot-info \
+      -eltorito-alt-boot \
+      -e '--interval:appended_partition_2:all::' \
+      -no-emul-boot \
+      -o "${DESTINATION_ISO}" "${ISO_FILESYSTEM_DIR}"
+  else
+    xorriso -as mkisofs \
+      -r -V "ubuntu-autoinstall-${TODAY}" -J \
+      -b isolinux/isolinux.bin \
+      -c isolinux/boot.cat \
+      -no-emul-boot \
+      -boot-load-size 4 \
+      -isohybrid-mbr /usr/lib/ISOLINUX/isohdpfx.bin \
+      -boot-info-table \
+      -input-charset utf-8 \
+      -eltorito-alt-boot \
+      -e boot/grub/efi.img \
+      -no-emul-boot \
+      -isohybrid-gpt-basdat -o "${DESTINATION_ISO}" "${ISO_FILESYSTEM_DIR}"
+  fi
 }
 
 args "$@"
